misc: [wip] Disable C++ features on shared mode
- disable C++ features (profiling and checkpointing) on shared mode
- remote unnecesary boost including
