+++ src/checkpoint/cpt_env.c
@@ -1,0 +1,9 @@
+#include <stddef.h>
+
+char *output_base_dir = NULL;
+char *config_name = NULL;
+char *workload_name = NULL;
+char *simpoints_dir = NULL;
+int cpt_id = -1;
+char *cpt_file = NULL;
+char *restorer = NULL;
