--- src/checkpoint/cpt_env.cpp
@@ -1,7 +1,0 @@
-char *output_base_dir = nullptr;
-char *config_name = nullptr;
-char *workload_name = nullptr;
-char *simpoints_dir = nullptr;
-int cpt_id = -1;
-char *cpt_file = nullptr;
-char *restorer = nullptr;
