+++ src/checkpoint/profiling.c
@@ -6,3 +6,8 @@ bool checkpoint_restoring = false;
 uint64_t checkpoint_interval = 0;
 
 bool profiling_started = false;
+
+#ifdef CONFIG_SHARE
+// empty definition on share
+void simpoint_profiling(uint64_t pc, bool is_control, uint64_t abs_instr_count) {}
+#endif
