+++ src/isa/riscv64/init.c
@@ -53,10 +53,14 @@ void init_isa() {
   vtype->val = 0; // actually should be 1 << 63 (set vill bit to forbidd)
 #endif // CONFIG_RVV_010
 
+#ifndef CONFIG_SHARE
   extern char *cpt_file;
   if (cpt_file == NULL) {
     memcpy(guest_to_host(RESET_VECTOR), img, sizeof(img));
   }
+#else
+  memcpy(guest_to_host(RESET_VECTOR), img, sizeof(img));
+#endif
 
   IFNDEF(CONFIG_SHARE, init_clint());
   IFDEF(CONFIG_SHARE, init_device());
